FunctionApp/requirements.txt
azure-functions
azure-storage-blob
azure-mgmt-containerinstance
azure-identity